Linear Algebra Models

LAModels is a module that provides a minimum interface needed by different linear algebra problems. This interface includes abstract types to specify the type of problem being considered, and it includes test to ensure that any subtypes of the abstract types contain the minimum necessary interface.

Abstract Types

Main.LAModels.LinearSystemsProblem — Type

LinearSystemsProblem

Abstract type to model a consistent linear systems problem,

\[Ax = b,\]

where $x$ is unknown.

Main.LAModels.LeastSquaresProblem — Type

LeastSquaresProblem

Abstract type to model a least squares problem,

\[\min_{x} \Vert Ax - b \Vert_2^2.\]

Subspace Approximation Problems

Subspace approximation problems will also have an abstract type and corresponding tests as this suite of packages evolves.

Macros for Testing

Main.LAModels.@test_linear_systems_problem — Macro

@test_linear_systems_problem(type)

Macro for implementing tests to verify that a sub-type of a LinearSystemsProblem has fields A::Matrix{Float64}, b::Vector{Float64}, sol::Vector{Float64}, and init::Vector{Float64} to ensure a common interface.

Main.LAModels.@test_least_squares_problem — Macro

@test_least_squares_problem(type)

Macro for implementing tests to verify that a sub-type of a LeastSquaresProblem has fields A::Matrix{Float64}, b::Vector{Float64}, sol::Vector{Float64}, and init::Vector{Float64} to ensure a common interface.